\name{Endosymbionts_GeneCalls}
\alias{Endosymbionts_GeneCalls}
\docType{data}
\title{
Example genecalls
}
\description{
A named list of \code{DataFrame}s.
}
\usage{data("Endosymbionts_GeneCalls")}
\format{
A named list.
}
\details{
Example genecalls.
}
\examples{
data(Endosymbionts_GeneCalls)
}
\keyword{datasets}
