\name{Endosymbionts_LinkedFeatures}
\alias{Endosymbionts_LinkedFeatures}
\docType{data}
\title{
Example synteny links
}
\description{
An object of class \code{LinkedPairs}.
}
\usage{data("Endosymbionts_LinkedFeatures")}
\format{
An object of class \code{LinkedPairs}.
}
\details{
An object of class \code{LinkedPairs}.
}
\examples{
data(Endosymbionts_LinkedFeatures)
}
\keyword{datasets}
