\name{Endosymbionts_Pairs01}
\alias{Endosymbionts_Pairs01}
\docType{data}
\title{
Example predicted pairs
}
\description{
An object of class \code{PairSummaries}.
}
\usage{data("Endosymbionts_Pairs01")}
\format{
An object of class \code{PairSummaries}.
}
\details{
An object of class \code{PairSummaries}.
}
\examples{
data(Endosymbionts_Pairs01)
}
\keyword{datasets}
