\name{Endosymbionts_Pairs02}
\alias{Endosymbionts_Pairs02}
\docType{data}
\title{
Example predicted pairs
}
\description{
An object of class \code{PairSummaries} where blocks have been expanded.
}
\usage{data("Endosymbionts_Pairs02")}
\value{
An object of class \code{PairSummaries}.
}
\details{
An object of class \code{PairSummaries}.
}
\examples{
data(Endosymbionts_Pairs02)
}
\keyword{datasets}
