\name{Endosymbionts_Sets}
\alias{Endosymbionts_Sets}
\docType{data}
\title{
A list of disjoint sets.
}
\description{
A named list of disjoint sets representing hypothetical COGs.
}
\usage{data("Endosymbionts_Sets")}
\format{
A named list of disjoint sets representing hypothetical COGs.
}
\details{
A named list of disjoint sets representing hypothetical COGs.
}
\examples{
data(Endosymbionts_Sets)
}
\keyword{datasets}