\name{Endosymbionts_Synteny}
\alias{Endosymbionts_Synteny}
\docType{data}
\title{
A synteny object
}
\description{
An object of class \code{Synteny}.
}
\usage{data("Endosymbionts_Synteny")}
\value{
An object of class \code{Synteny}.
}
\details{
An object of class \code{Synteny}.
}
\examples{
data(Endosymbionts_Synteny)
}
\keyword{datasets}
