\name{FeaturesFromDF}
\alias{FeaturesFromDF}
\title{
  Return a DNAStringSet of features
}
\description{
  Given a DNAStringSet of genomic sequences, and a \code{DataFrame} created by \code{\link{SquaregffBy}} containing the corresponding genecalls for the sequence, return a \code{DNAStringSet} representing the features in the \code{DataFrame}.
}
\usage{
  FeaturesFromDF(Genome,
                 GeneCalls)
}
\arguments{
  \item{Genome}{
    A \code{DNAStringSet}.
  }
  \item{GeneCalls}{
    A \code{DataFrame} created by \code{\link{SquaregffBy}} containing information on feature locations in the \code{DNAStringSet} supplied to \code{Genome}.
  }
}
\details{
  Given a \code{DataFrame} with the following feature columns:
  \itemize{
    \item \code{Index} - integers indicating specific \code{DNAString}s within the \code{DNAStringSet}
    \item \code{Strand} - integers (limited to 0 and 1) specifying the strandedness of the feature, 0 for the + strand, and 1 for the - strand
    \item \code{Range} - an \code{IRangesList} populated by subfeature boundaries the given feature; either the feature bounds themselves for non-coding features, or the CDS bounds (these can be out of phase!) for coding features
  }
  \code{DataFrame}s with these columns are auto-generated by both \code{\link{SquaregffBy}} and the soon to be defunct function \code{\link{gffToDataFrame}} (this function does not properly handle complex eukaryotic genecalls! \code{SquaregffBy} does!), but can be generated by users if necessary. With both this \code{DataFrame} and a \code{DNAStringSet} containing the genomic sequences associated with the genecalls, this function returns a \code{DNAStringSet} of the nucleotide sequences for the features present in the \code{DataFrame}. Features with subfeatures are stitched together, and features on the negative strand are \code{\link{reverseComplement}}ed.
}
\value{
  A \code{DNAStringSet} of features described in a given set of genomic sequences.
}
\author{
  Nicholas Cooley \email{Nicholas.Cooley@ul.ie}
}

\seealso{
  \itemize{
    \item \code{\link{SquaregffBy}}
    \item \code{\link{SummarizePairs}}
    \item \code{\link{NucleotideOverlap}}
    \item \code{\link{FrameDownward}}
  }
}
\examples{
file01 <- system.file("extdata",
                      "Endosymbionts_v05a.sqlite",
                      package = "SynExtend")
load(file = system.file("data",
                        "Endosymbionts_GeneCalls.RData",
                        package = "SynExtend"))
seqs <- SearchDB(dbFile = file01,
                 identifier = "1",
                 nameBy = "description")
res <- FeaturesFromDF(Genome = seqs,
                      GeneCalls = Endosymbionts_GeneCalls[[1]])
}
