\name{genecalls}
\alias{genecalls}
\docType{data}
\title{
Example genecall data
}
\description{
  A list of DataFrames.
}
\usage{data("genecalls", package = "SynExtend")}
\format{
  A list of \code{DataFrame}s.
}
\details{
  A list of genecalls generated by the \code{extdata.R} script contained in SynExtend's \code{inst/scripts} folder. This object contains succinct data for runnable function examples.
}
\examples{
data("genecalls", package = "SynExtend")
}
\keyword{datasets}
