\name{init_pairs}
\alias{init_pairs}
\docType{data}
\title{
Example genecall data
}
\description{
  An object of class \code{LinkedPairs}
}
\usage{data("init_pairs", package = "SynExtend")}
\format{
  An object of class \code{LinkedPairs}.
}
\details{
  An object of class \code{PairSummaries} generated by the \code{extdata.R} script contained in SynExtend's \code{inst/scripts} folder. This object contains succinct data for runnable function examples.
}
\examples{
data("init_pairs", package = "SynExtend")
}
\keyword{datasets}
