\name{linked_features}
\alias{linked_features}
\docType{data}
\title{
Example genecall data
}
\description{
  An object of class \code{LinkedPairs}
}
\usage{data("linked_features", package = "SynExtend")}
\format{
  An object of class \code{LinkedPairs}.
}
\details{
  An object of class \code{LinkedPairs} generated by the \code{extdata.R} script contained in SynExtend's \code{inst/scripts} folder. This object contains succinct data for runnable function examples.
}
\examples{
data("linked_features", package = "SynExtend")
}
\keyword{datasets}
