\name{Nitrosocosmicus_GeneCalls}
\alias{Nitrosocosmicus_GeneCalls}
\docType{data}
\title{
Example genecalls
}
\description{
A named list of \code{DataFrame}s.
}
\usage{data("Nitrosocosmicus_GeneCalls")}
\format{
A named list.
}
\details{
Example genecalls.
}
\examples{
data(Nitrosocosmicus_GeneCalls)
}
\keyword{datasets}
