\name{Nitrosocosmicus_LinkedFeatures}
\alias{Nitrosocosmicus_LinkedFeatures}
\docType{data}
\title{
Example synteny links
}
\description{
An object of class \code{LinkedPairs}.
}
\usage{data("Nitrosocosmicus_LinkedFeatures")}
\format{
An object of class \code{LinkedPairs}.
}
\details{
An object of class \code{LinkedPairs}.
}
\examples{
data(Nitrosocosmicus_LinkedFeatures)
}
\keyword{datasets}