\name{Nitrosocosmicus_Pairs02}
\alias{Nitrosocosmicus_Pairs02}
\docType{data}
\title{
Example predicted pairs
}
\description{
An object of class \code{PairSummaries} where blocks have been expanded.
}
\usage{data("Nitrosocosmicus_Pairs02")}
\format{
An object of class \code{PairSummaries}.
}
\details{
An object of class \code{PairSummaries}.
}
\examples{
data(Nitrosocosmicus_Pairs02)
}
\keyword{datasets}