\name{Nitrosocosmicus_Pairs03}
\alias{Nitrosocosmicus_Pairs03}
\docType{data}
\title{
Example predicted pairs
}
\description{
An object of class \code{PairSummaries} where blocks have been expanded and competitors have been rejected.
}
\usage{data("Nitrosocosmicus_Pairs03")}
\format{
An object of class \code{PairSummaries}.
}
\details{
An object of class \code{PairSummaries}.
}
\examples{
data(Nitrosocosmicus_Pairs03)
}
\keyword{datasets}