\name{Nitrosocosmicus_Sets}
\alias{Nitrosocosmicus_Sets}
\docType{data}
\title{
A list of disjoint sets.
}
\description{
A named list of disjoint sets representing hypothetical COGs.
}
\usage{data("Nitrosocosmicus_Sets")}
\format{
A named list of disjoint sets representing hypothetical COGs.
}
\details{
A named list of disjoint sets representing hypothetical COGs.
}
\examples{
data(Nitrosocosmicus_Sets)
}
\keyword{datasets}