\name{Nitrosocosmicus_Synteny}
\alias{Nitrosocosmicus_Synteny}
\docType{data}
\title{
A synteny object
}
\description{
An object of class \code{Synteny}.
}
\usage{data("Nitrosocosmicus_Synteny")}
\format{
An object of class \code{Synteny}.
}
\details{
An object of class \code{Synteny}.
}
\examples{
data(Nitrosocosmicus_Synteny)
}
\keyword{datasets}