\name{NucleotideOverlap}
\alias{NucleotideOverlap}
\title{
Tabulating Features Linked by Syntenic Hits
}
\description{
A function for concisely tabulating where genomic features are connected by syntenic hits.
}
\usage{
NucleotideOverlap(SyntenyObject,
                  GeneCalls,
                  LimitIndex = FALSE,
                  AcceptContigNames = TRUE,
                  Verbose = FALSE)
}
\arguments{
  \item{SyntenyObject}{
An object of class \code{Synteny} built from the \code{FindSynteny} function in the package \code{DECIPHER}.
}
  \item{GeneCalls}{
A named list of \code{DataFrame}s built from \code{SquaregffBy}, objects of class \code{GRanges} imported from \code{rtracklayer::import}, or objects of class \code{Genes} created from the \code{DECIPHER} function \code{FindGenes}. \code{DataFrame}s built by \code{SquaregffBy} can be used directly, while \code{GRanges} objects may also be used with limited functionality. Objects of class \code{Genes} generated by \code{FindGenes} function equivalently to those produced by \code{SquaregffBy}.
}
  \item{LimitIndex}{
Logical indicating whether to limit which indices in a synteny object to query. \code{FALSE} by default, when \code{TRUE} only the first sequence in all selected identifiers will be used. \code{LimitIndex} can be used to skip analysis of plasmids, or solely query a single chromosome.
}
  \item{AcceptContigNames}{
Match names of contigs between gene calls object and synteny object. Where relevant, the first white space and everything following are removed from contig names. If ``TRUE'', NucleotideOverlap assumes that the contigs at each position in the synteny object and ``GeneCalls'' object are in the same order. Is automatically set to TRUE when ``GeneCalls'' are of class ``GRanges''.
}
  \item{Verbose}{
Logical indicating whether or not to display a progress bar and print the time difference upon completion.
}
}
\details{
  Builds a matrix of lists that contain information about linked pairs of genomic features.
}
\value{
  A matrix of lists with location information about syntenic hits that link features. The upper triangle is populated by information about the hits themselves, while the lower triangle is populated by summary statistics of the linked features.
}
\author{
  Nicholas Cooley \email{Nicholas.Cooley@ul.ie}
}

\seealso{
  \itemize{
    \item \code{\link{FindSynteny}}
  }
}
\examples{
data("genecalls", package = "SynExtend")
data("syn", package = "SynExtend")

x <- NucleotideOverlap(SyntenyObject = syn,
                       GeneCalls = genecalls,
                       LimitIndex = FALSE,
                       Verbose = TRUE)
}
