\name{syn}
\alias{syn}
\docType{data}
\title{
Example genecall data
}
\description{
  An object of class \code{Synteny}
}
\usage{data("syn", package = "SynExtend")}
\format{
  An object of class \code{Synteny}.
}
\details{
  An object of class \code{Synteny} generated by the \code{extdata.R} script contained in SynExtend's \code{inst/scripts} folder. This object contains succinct data for runnable function examples.
}
\examples{
data("syn", package = "SynExtend")
}
\keyword{datasets}
